safari-600.7-branch/LayoutTests/inspector-obsolete/console/console-xpath-expected.txt

CONSOLE MESSAGE: line 1: complete
Tests $x for iterator and non-iterator types.

$x('42')
$x('name(/html)')
$x('not(42)')
$x('/html/body/p').length
42
"html"
false
1
complete
